refactor(examples): focus workflow on orchestrator
