+++ examples/custom_backend/workflow_hello_world/workflow.py
@@ -1,31 +1,16 @@
 # SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
 
-"""Author and bind the declarative workflow."""
-
-from typing import Any
+"""Author and compile the declarative Hello World workflow."""
 
 from dynamo.workflow import (
     DeploymentSpec,
     ExecutionPlan,
     ValueSpec,
     Workflow,
-    WorkflowOrchestrator,
     compile_workflow,
 )
-from examples.custom_backend.workflow_hello_world.dynamo.stages import (
+from examples.custom_backend.workflow_hello_world.stages import (
     HelloStage,
     MergeStage,
     WorldStage,
@@ -57,11 +42,4 @@ def compile_remote_workflow() -> ExecutionPlan:
     return compile_workflow(
         define_workflow(),
         DeploymentSpec.remote(**ENDPOINTS),
-    )
-
-
-async def provide_workflow(runtime: Any) -> WorkflowOrchestrator:
-    return await WorkflowOrchestrator.bind(
-        compile_remote_workflow(),
-        runtime=runtime,
     )
